pyproject.toml
[project]
name = "memu-server"
version = "0.1.0"
description = "backend wrapper for memU"
readme = "README.md"
requires-python = ">=3.13"

dependencies = [
    # Web Framework
    "fastapi[standard]>=0.122.0",
    "uvicorn[standard]>=0.35.0",

    # Memory Service (includes sqlmodel, openai, pendulum, etc.)
    "memu-py[postgres]>=1.2.0",

    # Database driver (needed for connection URL construction)
    "psycopg[binary,pool]>=3.2.9",

    # Workflow Engine
    "temporalio==1.16.0",

    # Production server
    "gunicorn>=23.0.0",

    # Configuration & Utils
    "pydantic-settings>=2.10.1",
    "python-dotenv>=1.0.0",
]

[dependency-groups]
dev = [
    "pytest>=7.2.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=6.0.0",
    "httpx>=0.27.0",
    "ruff>=0.11.0",
    "pre-commit>=4.0.0",
    "bandit[toml]>=1.8.0",
    "pylint>=3.0.0",
    "mypy>=1.8.0",
    "deptry>=0.16.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app", "config"]

[tool.ruff]
line-length = 120
target-version = "py313"

[tool.ruff.lint]
# Enable import sorting checks
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort (import sorting)
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
]
ignore = []

[tool.ruff.lint.isort]
known-first-party = ["app", "config"]
section-order = ["future", "standard-library", "third-party", "first-party", "local-folder"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.coverage.run]
source = ["app", "config"]
omit = [
    "*/tests/*",
    "*/__pycache__/*",
    "*/site-packages/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]

[tool.bandit]
exclude_dirs = [".venv", "venv", "tests", ".git", "__pycache__"]
# B105: hardcoded_password_string
# B106: hardcoded_password_funcarg
# B107: hardcoded_password_default
skips = []

[tool.pylint.messages_control]
disable = ["C0111", "C0103"]  # missing-docstring, invalid-name
enable = ["W0102", "W0212", "W0611"]  # dangerous-default-value, protected-access, unused-import

[tool.deptry]
# Ignore dependencies that are planned for future use but not yet implemented
[tool.deptry.per_rule_ignores]
DEP002 = [
    "uvicorn",           # Will be used for production deployment
    "psycopg",           # PostgreSQL adapter used by memu-py[postgres] and alembic
    "temporalio",        # Planned for workflow orchestration
    "python-dotenv",     # Environment variables (loaded by pydantic-settings)
]
DEP003 = [
    "app",               # Project's own package, not an external dependency
    "config",            # Project's own config package
    "pydantic",          # Re-exported by pydantic-settings
]

[tool.mypy]
python_version = "3.13"
files = ["app", "config"]
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false  # Gradually enable
ignore_missing_imports = true
